+++ src/App.tsx
@@ -5,10 +5,8 @@ import EquipmentPage from "./components/Resources/EquipmentPage";
 import { Routes, Route, Navigate } from "react-router-dom";
 import "./App.css";
 import ScheduleComponent from "./components/Schedule/ScheduleComponent";
-import BladeProjectPage from "./components/Projects/BladeProjectPage";
 import BTPage from "./components/BladeTask/BladeTaskPage";
-import BladeProjectPageWithSchedule from "./components/Projects/BladeProjectPageSchedule";
-import BladeProjectPageWithScheduleViewEdit from "./components/Projects/BladeProjectPageEditView";
+import BladeProjectPage from "./components/Projects/BladeProjectPage";
 
 
 import { EditModeProvider } from "./EditModeContext";
@@ -26,7 +24,7 @@ function App() {
                         <Route path="/" element={<ScheduleComponent/>} />
                         <Route
                             path="/projects"
-                            element={<BladeProjectPageWithScheduleViewEdit />}
+                            element={<BladeProjectPage />}
                         />
                         <Route path="/bladetask" element={<BTPage />} />
                         <Route path="/resources" element={<EquipmentPage />} />
