+++ src/components/TableLogic/TableLogic.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React from "react";
+import React,{useState}from "react";
 import StyledButton from "../ui/styledButton";
 import { useCollapse } from "react-collapsed";
 import { Input } from "../ui/input";
@@ -69,14 +69,16 @@ export function TableLogic<TData, TValue>({
 }: DataTableProps<TData, TValue>) {
 
     //List of useState hooks that are used to set the state of the table.
-    const [sorting, setSorting] = React.useState<SortingState>([]);
-    const [columnVisibility, setColumnVisibility] = React.useState({});
+    const [sorting, setSorting] = useState<SortingState>([]);
+    const [columnVisibility, setColumnVisibility] = useState({});
     const [columnFilters, setColumnFilters] =
-        React.useState<ColumnFiltersState>([]);
-    const [filtering, setFiltering] = React.useState("");
-    const [expanded, setExpanded] = React.useState<ExpandedState>({});
-    const [inputValue, setInputValue] = React.useState("");
+        useState<ColumnFiltersState>([]);
+    const [filtering, setFiltering] = useState("");
+    const [expanded, setExpanded] = useState<ExpandedState>({});
+    const [inputValue, setInputValue] = useState("");
     const { getToggleProps } = useCollapse();
+    const [showPasswordPrompt, setShowPasswordPrompt] = useState(false);
+
 
     /**  
      * The useReactTable hook is used to create a table .
@@ -132,7 +134,7 @@ export function TableLogic<TData, TValue>({
                     onChange={(e) => {
                         setFiltering(e.target.value);
                     }}
-                    className="max-w-sm"
+                    className="w-80 r"
                 />
                 <div>
                     {/* popover for changing column visibility*/}
@@ -193,7 +195,7 @@ export function TableLogic<TData, TValue>({
                     </StyledButton>
                 </div>
                 <div>
-                    <SwitchComponent/>
+                    <SwitchComponent setShowPasswordPrompt={setShowPasswordPrompt}/>
                 </div>
             </div>
